refactor: move calendar feature to features/calendar/
