+++ app/calendar/page.tsx
@@ -4,15 +4,15 @@
 import { useState, useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import { CalendarIcon, LayoutList, GanttChartSquare, Plus } from 'lucide-react'
-import { Button } from '@/components/ui/button'
-import { SchoolSidebar } from '@/components/calendar/SchoolSidebar'
-import { TimelineView } from '@/components/calendar/TimelineView'
-import { GanttView } from '@/components/calendar/GanttView'
-import { ConflictBanner } from '@/components/calendar/ConflictBanner'
-import { AddScheduleModal } from '@/components/calendar/AddScheduleModal'
-import { toBookmark } from '@/components/calendar/types'
-import type { Bookmark, ViewType, UserScheduleRow } from '@/components/calendar/types'
-import { createBrowserClient } from '@/lib/supabase/browser'
+import { Button } from '@/shared/ui/button'
+import { SchoolSidebar } from '@/features/calendar/components/SchoolSidebar'
+import { TimelineView } from '@/features/calendar/components/TimelineView'
+import { GanttView } from '@/features/calendar/components/GanttView'
+import { ConflictBanner } from '@/features/calendar/components/ConflictBanner'
+import { AddScheduleModal } from '@/features/calendar/components/AddScheduleModal'
+import { toBookmark } from '@/features/calendar/types'
+import type { Bookmark, ViewType, UserScheduleRow } from '@/features/calendar/types'
+import { createBrowserClient } from '@/shared/lib/supabase/browser'
 
 export default function CalendarPage() {
   const router = useRouter()
